new work

+++ libft/inc/libft.h
@@ -31,7 +31,7 @@ int		ft_isprint(int c);
 size_t	ft_strlen(const char *s);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
-int		ft_atoi(const char *nptr);
+long long		ft_atoi(const char *nptr);
 int		ft_strncmp(char *s1, char *s2, unsigned int n);
 int		ft_strcmp(const char *s1, const char *s2);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
